+++ tests/System.IO.Pipelines.Extensions.Tests/BufferPoolFacts.cs
@@ -2,7 +2,6 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using System.IO.Pipelines.Networking.Sockets.Internal;
 using Xunit;
 
 namespace System.IO.Pipelines.Tests
