+++ tests/System.IO.Pipelines.Extensions.Tests/SignalFacts.cs
